Split data types definition for dioxane/mock coin
into separate package. Split is very suboptimal since we havve to pass transtion
functions as well
