+++ nix/release.nix
@@ -39,17 +39,18 @@ let
     bls-signatures = callInternal hsPkgs "bls-signatures" ../bls-signatures {} "";
     hschain-crypto = callInternal hsPkgs "hschain" ../hschain-crypto {}
       (if useSodium then "-flibsodium" else "-f-libsodium");
-    hschain-crypto-bls = callInternal hsPkgs "hschain" ../hschain-crypto-bls    {} "";
-    hschain-quickcheck = callInternal hsPkgs "hschain" ../hschain-quickcheck    {} "";
-    hschain-types      = callInternal hsPkgs "hschain" ../hschain-types         {} "";
-    hschain-merkle     = callInternal hsPkgs "hschain" ../hschain-merkle        {} "";
-    hschain-net        = callInternal hsPkgs "hschain" ../hschain-net           {} "";
-    hschain-pow-func   = callInternal hsPkgs "hschain" ../proof-of-work         {} "";
-    hschain-PoW        = callInternal hsPkgs "hschain" ../hschain-PoW           {} "";
-    hschain            = callInternal hsPkgs "hschain" ../hschain               {} "";
-    hschain-examples   = callInternal hsPkgs "hschain" ../hschain-examples      {} "";
-    hschain-control    = callInternal hsPkgs "hschain" ../hschain-control       {} "";
-    serialise-cddl     = callInternal hsPkgs "hschain" ../serialise-cddl        {} "";
+    hschain-crypto-bls     = callInternal hsPkgs "hschain" ../hschain-crypto-bls     {} "";
+    hschain-quickcheck     = callInternal hsPkgs "hschain" ../hschain-quickcheck     {} "";
+    hschain-types          = callInternal hsPkgs "hschain" ../hschain-types          {} "";
+    hschain-merkle         = callInternal hsPkgs "hschain" ../hschain-merkle         {} "";
+    hschain-net            = callInternal hsPkgs "hschain" ../hschain-net            {} "";
+    hschain-pow-func       = callInternal hsPkgs "hschain" ../proof-of-work          {} "";
+    hschain-PoW            = callInternal hsPkgs "hschain" ../hschain-PoW            {} "";
+    hschain                = callInternal hsPkgs "hschain" ../hschain                {} "";
+    hschain-examples       = callInternal hsPkgs "hschain" ../hschain-examples       {} "";
+    hschain-examples-types = callInternal hsPkgs "hschain" ../hschain-examples-types {} "";
+    hschain-control        = callInternal hsPkgs "hschain" ../hschain-control        {} "";
+    serialise-cddl         = callInternal hsPkgs "hschain" ../serialise-cddl         {} "";
   };
   # Build internal package
   callInternal = hask: name: path: args: opts:
@@ -94,12 +95,14 @@ let
       hschain-PoW
       hschain
       hschain-examples
+      hschain-examples-types
     ];
     hschainPkgJs = p: with p; [
       hschain-crypto
       hschain-control
       hschain-types
       hschain-merkle
+      hschain-examples-types
     ];
     in
     {
